# Conflicts:
#	imgui.ini
#	src/ModelInstance.cpp
#	src/SceneManager.cpp
#	src/Shadowcaster.cpp
#	src/Viewport.cpp
#	src/main.cpp

+++ dat/shaders/g_buffer.frag
@@ -10,7 +10,7 @@ uniform sampler2D tex_spec;
 uniform sampler2D tex_norm;
 uniform sampler2D tex_emit;
 
-uniform uvec4 obj_id;
+uniform vec4 obj_id;
 
 layout (location = 0) out vec3 g_position;
 layout (location = 1) out vec3 g_normal;
@@ -28,5 +28,6 @@ void main() {
 	g_spec     = texture( tex_spec, uv_fs );
 	g_albedo   = texture( tex_diff, uv_fs );
 	g_emit     = texture( tex_emit, uv_fs );
+	// g_spec	   = obj_id; // här ifall jag vill se picking texturen 
 	g_picking  = obj_id;
 }
